Fix keyboard-layout being changed to US on RDP connection
Added the IgnoreRemoteKeyboardLayout Registry Key to prevent automatic changes to the keyboard layout.

As mentioned in (this article)[https://poweradm.com/ignoreremotekeyboardlayout-windows-rdp/] the rdp server adds and activates the US keyboard-layout if the keyboard-layout of the client does not match the servers. Since our client is on linux there are many cases where this is not given, so it is better to let the user choose a layout on windows.